+++ example/bounce/view/board.dart
@@ -5,7 +5,7 @@ class Board extends Surface {
 
   Board(CanvasElement canvas): super(canvas) {
     movablePieces = new MovablePieces(32,
-        distance: new Distance.from(canvas.width, canvas.height));
+        new Distance.from(canvas.width, canvas.height));
     movablePieces.randomInit();
     movablePieces.forEach((MovablePiece p) => p.shape = PieceShape.CIRCLE);
   }